+++ resources/views/siulomos/index.blade.php
@@ -32,12 +32,15 @@
                             <td>{{ $siuloma->stud_limitas }}</td>
                             <td>
                                 <a href="/siuloma/{{$siuloma->id}}" class="btn btn-primary">Detaliau</a>
-                                <a href="/siuloma/edit/{{$siuloma->id}}" class="btn btn-primary">Redaguoti</a>
-                                <a href="/siuloma/delete/{{$siuloma->id}}" class="btn btn-danger">Šalinti</a>
-                                <a href="/siuloma/approve/{{$siuloma->id}}" class="btn btn-success">Patvirtinti</a>
-{{--                                @if($tema->stud_limitas - $tema->pasirinkusieji > 0 && !auth()->user()->pasirinkta_tema)--}}
-{{--                                    <a href="/tema/choose/{{$tema->id}}" class="btn btn-success">Rinktis</a>   Galima rinktis tik atitikus salygas--}}
-{{--                                @endif--}}
+                                @if(auth()->user()->isAdmin() || $siuloma->user_id == auth()->user()->id)
+                                    <a href="/siuloma/edit/{{$siuloma->id}}" class="btn btn-primary">Redaguoti</a>
+                                @endif
+                                @if(auth()->user()->isAdmin() || $siuloma->user_id == auth()->user()->id)
+                                    <a href="/siuloma/delete/{{$siuloma->id}}" class="btn btn-danger">Šalinti</a>
+                                @endif
+                                @if(auth()->user()->isAdmin())
+                                    <a href="/siuloma/approve/{{$siuloma->id}}" class="btn btn-success">Patvirtinti</a>
+                                @endif
                             </td>
                         </tr>
                     @endforeach
